ReviewManager/Capabilities/requirements.txt
black==26.3.1
blessed==1.38.0
boto3==1.42.88
botocore==1.42.88
chalice==1.31.2
click==8.3.2
inquirer==2.10.1
isort==8.0.1
jmespath==1.1.0
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.6
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-editor==1.0.4
pytokens==0.4.1
PyYAML==6.0.3
readchar==4.2.2
s3transfer==0.16.0
six==1.17.0
urllib3==2.6.3
wcwidth==0.6.0
